+++ Changelog.md
@@ -4,6 +4,7 @@
 - **2.5.1-SNAPSHOT** 
    - Prepare for next dev cycle
    - Renamed native folder where the executables are to nativebin too, to prevent Java 9+ problems
+   - Allow to create a MultimediaObject with an URL instead of a file as conversion source
 - **2.5.0** 
    - Prepare for next dev cycle
    - Renamed build artifacts with native in the name to nativebin to avoid module conflicts in Java 9+